Switch from deprecated wmic to powershell CMI cmdlet
The long time deprecated wmic tool was dropped in Windows 11 WIP and it
should be dropped in the next release of Windows 10 and Windows 11.

CMI cmdlets require powershell 3.0 or later, so a Windows 7 user
may need to update to a newer powershell version.

The legacy VMI cmdlets were dropped with powershell 6.0, so they are not
a viable solution to keep compatibility with old (and EOL) Windows versions.

Issue and solutions analysis by @d3vv

closes #1217