Allow expressions involving max for concurrency and max_memory.
Examples:

--concurrency max-1            # leave 1 core for the OS
--concurrency min(4,max/2)     # reserve half of the cores for
                                 Fishtest with a maximum of 4

We compute the defaults for concurrency and max_memory in the configfile.

A freshly minted configfile now contains

concurrency = 1 if max==1 else min(3, max-1)
max_memory = max/2

If the expression for concurrency does not contain "max",
concurrency is limited to max_cpu_count-1. Otherwise it is limited to
max_cpu_count.