Earlier validation of --concurrency and --max_memory.
Also:

Put the actual values for concurrency and max_memory as comments
in the config file.

Also:

Rename max to MAX and inject the function max in the parser.

As a temporary measure, for backward compatibility "max", when
used as a variable, is automatically converted to "MAX",
both in the config file and on the command line. When "max" is
used as a variable on the command line, a deprecation warning
is given.