New shot at generating a stable default uuid_prefix for workers..
Currently the algorithm for the default uuid_prefix uses the MAC
address. But the MAC address is unstable on modern OSes. Moreover
uuid.getnode() seems to be buggy on a number of systems and simply
returns a random number. This needs ugly and fragile workarounds.

The new algorithm uses:

1) the machine_id (a uuid generated during system install);
2) the absolute path of the worker;
3) a random seed which is saved in the config file.

This algorithm can be defeated by cloning a cloud instance with
an installed worker.

This seems like a niche case. But if it turns out that it does happen
then we can work around it by detecting the resulting duplicate
prefix and generating a new seed.