Improve the purging PGNs script
The "pgns" collection inflated with old PGNs because: some devs's pattern
may stop the purging loop (eg pausing a run for some weeks), the purge
script was updated with delay with the changes to the run dictionary template.
The "pgns" collection was cleaned with a temporary custom script update.

Changes:
- stop the purging loop based on "last_update" instead than "start_time"
- purge the not finished runs if "last_updated" is older than 50 days
- use python3