Remove target=blank attribute for username link
If we do a sequence of
LiveElo -> username -> LiveElo -> username -> LiveElo -> username,
we end up with an infinite number of tabs.

All other links in fishtest pointing to https://tests.stockfishchess.org/
remain in the same page, giving the global impression that
https://tests.stockfishchess.org/ is a single window app
(there must be a technical term for that). This patch is
coherent with this.

closes https://github.com/glinscott/fishtest/pull/1438