Compute sri hashes of worker files and compare with master.
The sri hashes of the current worker files are stored in a
file "sri.txt". This file is updated each time the worker
runs, and it can also be updated manually with

python worker.py --only_config

CI is used to enforce that the file sri.txt matches the worker files
during pushes or pull requests.

The worker will also download the master sri.txt (currently not existing)

https://raw.githubusercontent.com/glinscott/fishtest/master/worker/sri.txt

and check if the hashes in that file coincide with the sri hashes
of the current worker. If not then a message is printed on the terminal.
A tainted worker is not stopped but the flag worker_info['modified'] is
set to True. In a subsequent PR the server will use this information
in the event log messages (for debugging purposes).

In order to fix the race condition between the update of the master
worker on GitHub and the update of the Fishtest server (which pulls its
code from GitHub), the file sri.txt contains the version number of the
worker. We ignore the master sri.txt if its version is higher
than the version of the running worker.