Asynchronous network upload
To avoid duplicated netowrk entries in the DB
was forced a synchronous upload, see:
https://github.com/glinscott/fishtest/pull/757

An easy way to enable to avoid duplicate entries is to
look up the network in the DB just before adding the entry.

This PR speed up the upload process and reduces possible race conditions
checking for duplicate files on the file server as well, raising
the appropriate HTTP errors if necessary.

Other changes:
 - use a fixed path for the POST (usually the {id} is used in PUT and
   DELETE)
 - use a multipart form POST to upload the network.
   In this way the file server can use an UploadFile object
   that uses in a smart way the RAM for big data

Requirements:
 - update of the file server
 - adding a "/" in the fishtest.upload