Reduce "LTC only" thread threshold
Several workers in the last month between 30 and 32 threads have fallen afoul of the well-established cutechess bugs in many-threaded contexts. Altho most such workers run fine, we currently have no server-side filtering of garbage results, nor any worker-side workaround, so this is the best we've got. Update it now to reduce future garbage with the least current effort.

As stated, most workers in the 29-32 thread run fine, but those that don't are a significant problem. To the best of my knowledge, all workers at 28 or less have run fine so far. It's certainly easy to adjust this again in the future.