Don't acquire rundb.run_cache_lock when generating webpages
This lock is acquired every time we retrieve an individual run from the DB, so it's kind of an important lock.

On current master, while rendering the homepage, we might acquire and release this core lock dozens of times, once for each run,
and while acquired we prevent other threads from getting runs from the DB.

But we only do this while retrieving the run results, but we don't even need the lock to retrieve the results! It's just that
`rundb.get_results(run)` is used by both internal DB-updating code and by the webpage rendering code, and the former may need
to save its work with the lock, while rendering webpages is very much *not* an operation that writes to the DB -- no need to lock.

So instead we stick to "read-only mode", so to speak, which already exists, when using `rundb.get_results(run)` to render webpages.
And for the homepage in particular, we save dozens of acquire-release cycles, which could create myriad conflicts with other threads
... but no longer.