Increase cache time for user creditials
increases the time to 2min, which should halve the time spent in authetication.
Authentication happes quite often, i.e. each worker connects every minute,
and is one of the more expensive calls for the critical pserve.

The PR also clears the cache on writes to the database,
so that e.g. a password change takes effect immediately.

It appears possible to just use the cache independent of the age of the entry,
as long as the cache is cleared on every db write. This has not yet been implemented,
but seems straightforward.