Adjust batch size for SPRT tests.
an earlier commit made the batch size for SPRT test 32. This is a good number of single-threaded STC tests,
but not for multi-threaded LTC tests. Tests running 60+0.6@8t take too long on an 8core worker to update results.
The following patch adjust the number of games, keeping it the same for single threaded STC, but scaling down with TC and threading.