Regularly purge stale notifications.
Stale notifications are notifications for finished runs.
They are harmless but they needlessly load the server (forever with
the current implementation).

Stale notifications can for example be created by subscribing to a
finished run in a browser tab that was not refreshed.

Purging will happen every 5.5 hours on average and uses the
active_runs api to detect finished but still subscribed runs.