Wrap saving/loading objects from localstorage.
This makes the code in notifications.js a bit more readable.

Functions:

load_object(key)
save_object(key, value)

The functions JSON.stringify() and JSON.parse() are applied
under the hood.

When saving to localstorage, keys are given a prefix __fishtest__.
This makes upgrading easier since we can just change the prefix
to something else, e.g. __fishtest_v2__.

Also: fix bug. The username reported for stopped/deleted runs
was the user doing the stopping/deleting and not the owner
of the run.

After this PR all tabs should be reloaded.