Standardize usage of "last_updated" name
In the codebase, the key "last_updated" is consistently used as
a datetime object. However, in "delta_update_users.py",
it was used as a string representation.

This change standardizes the usage of "last_updated" by ensuring that
it is always used as a datetime object. As a result, it is necessary
to run "delta_update_users.py" in full scan mode to rebuild the
"users_cache" and "top_month" collections with the new key names.