Reduce ITP with number of active tests per user
given some excessive number of active tests by some users, this patch reduces itp based on the number of active tests.

The formula used is `itp *= 36.0 / (36.0 + count * count)`

This reduces slowly the itp for a number of tests smaller than 6, and rapidly afterwards.
For 3 active tests, the itp is 80%, for 6 active tests, the itp is 50%, for 15 active tests it is down to 13%.

For a maximum share of fishtest... have 6 tests in queue.