Prefer the stateless approach over the stateful one for submitting tests.
After experimenting with the approach proposed, it appears that the new approach break many workflows that submitter adapted for, which cannot be trivially fixed due to the stateful nature of this approach.

an example of that is the test run page requires another user interaction to reflect a new change on the branch after fetching first time occurred.