First shot at worker blocking.
Approvers and owners can block/unblock.

A blocked worker gets an error message which can be customized
by the blocker.

The html is atrocious but I have no time to improve it.
Hopefully an expert will improve it.

NOTE: An index must be generated on the workers collection.

Currently the workerdb keeps growing. In the future stale entries
should be purged regularly.