Implement efficient PGN upload and bulk download
The previous worker code posted the PGN to the server in a
zlib-compressed format. This required the server to decompress
the data before serving the PGN as text, resulting in significant CPU
and network overhead. Furthermore, a bulk download of all PGNs from a run
necessitated additional CPU overhead for compression.

The updated worker code posts a gzip file containing the PGN text
to the server, leading to more efficient handling on the server side:
- for single PGN downloads, the server simply serves
  the gzipped PGN retrieved from MongoDB. All modern browsers
  decompress the gzip file client side, so the user will get the
  pgn as text file. nginx can be configured to decompress on the fly
  the gzip content for a client that is not declaring his compression
  capability.
- for bulk PGN downloads, the server writes a tar file
  containing all the gzipped PGNs retrieved from MongoDB
  ```
  wget https://test.stockfishchess.org/api/run_pgns/<run_id>.pgns.tar
  ```
- gzip module uses zlib, no change for the compression ratio.

@peregrineshahin wrote the bulk download code
in https://github.com/official-stockfish/fishtest/pull/1804

Retire "pgn_100" api.

closes https://github.com/official-stockfish/fishtest/issues/1801