Implement efficient PGN upload and bulk download
The previous worker code posted the PGN to the server in an encoded and
zlib-compressed format. This required the server to decompress and decode
the data before serving the PGN as text, resulting in significant CPU
and network overhead. Furthermore, a bulk download of all PGNs from a run
necessitated additional CPU overhead for compression.

The updated worker code posts a zip file containing the PGN text
to the server, leading to more efficient handling on the server side:
- for single PGN downloads, the server simply serves
  the zipped PGN retrieved from MongoDB
- for bulk PGN downloads, the server writes a zip file in store mode
  containing all the zipped PGNs retrieved from MongoDB

@peregrineshahin wrote the bulk download code
in https://github.com/official-stockfish/fishtest/pull/1804

closes https://github.com/official-stockfish/fishtest/issues/1801