Small cleanups in the run db.
(1) Remove the field "results_stale" since now the results are
updated incrementally. So they should never be stale.

(2) Retire get_results() and replace it by compute_results()
which has no side effects. The latter is called in two places:
(a) in purge_run() (b) after a run is finished to verify the
incrementally updated results.

(3) Set the "workers" and "cores" fields equal to zero when
the run is finished.