Introduce generics in validation.
We can now write

"groups" : list[str]

in the schema for the userdb.

Other changes:

Put the validation code in a separate file.
Refactor the validation code so that it is easier to work with.
Introduce an email type.
Introduce a regex type (not yet used).
Create unit tests for the added code.