Introduce a schema for the rundb.
Validation is done after the creation of runs and before writing
finished runs to disk. If a finished run fails validation then
this is logged but the run is otherwise treated normally.

No attempt has been made to make the schema as tight as possible.
For example most numbers should be positive but this constraint
is ignored by the schema. The main purpose of the schema is to
document the structure of the rundb. The reason it is also used
for validation in some cases is to ensure that if the
structure of the rundb changes then the schema is updated to
reflect this change.

For a description of the schema format see the README.md file
of https://github.com/vdbergh/vtjson

The validation code is included in this PR but it can be
installed as an independent package via

pip install vtjson

Validating a run with 500 tasks (which is a lot these days)
takes 12 ms on an AWS t2.micro using Python 3.11.