Shut down the worker quickly after receiving SIGINT during make.
To test this PR (on Linux, see below) one may run the worker
in one terminal and during compiling send a SIGINT or SIGTERM
to the python process from another terminal (e.g. via
kill <PID-python>). A worker running master will only stop
when the make command finishes. If the worker runs this PR
then it will stop immediately and clean up.

I assume this PR will be useful if the worker is started via
a service manager like systemd which typically stops a
service by sending a SIGTERM to the main process. I have
not tested this however.

This PR currently has no effect on Windows as the mechanism
for terminating processes is different. Probably there is
also no need for this PR on that OS.