Convenient link to blocked workers
Currently, there is no convenient way to get to the particular worker page to block/unblock after misbehaving workers are spotted in test_view Tasks.
You would need to write the URL manually or search for that worker on the home page.

Make the approver-only anchor redirect to the worker management rather than the user page as the user page is irrelevant after we can block only the worker.