Don't hide "non-tested" nets for logged out users
there is no reason to hide them on the first place, and practice shows it's not the ideal thing let alone the net is still available so this is a fake frontend guard.

A follow up fix should be, website server changes after dual net, since the master net is not marked as tested for now.