Introduce log_message action.
This may be used to give internal errors in the logic of
Fishtest maximum visibility.

This action is currently used in two places:

- if a finished run does not validate;

- if check_results finds a discrepancy between the incrementally
updated results and the computed results in a finished run.

Currently both errors are logged but otherwise they are treated silently.

Another application in the future might be to log certain
worker errors which happen before a task is requested (e.g.
failure to be able to run cutechess-cli).

While at it, replace all occurences of _validate(), which returns
an error string on non-validation, by the more Pythonic
validate(), which throws an exception instead.