View actual git diff in tests_view
This request has been pending for years, since GitHub does not provide the diffs with two dots in the API we are not able to actually see the actual diff in test view, this replaces the request to GitHub API with three dots to a local algorithm with the help of jsDiff library to find the actual diff between files.

How can we continue from there:
make tabs for the diff section such that we can see the actual diff with official master without going to GitHub.

online tool made for testing at: https://githubdifftesting.netlify.app/