Avoid timeouts for GitHub diffs
This allows for users to use their own limited GitHub tokens to avoid exceeding the GitHub rate limit when viewing github diffs (mainly useful to approvers), which is set at 60 requests per hour for unauthenticated users. By using this token, you can surpass the standard rate limit, extending it beyond the 5000 requests per hour limit. You can find more information about GitHub's rate limits here.

It's important to note that this token will not be stored in a database but rather in your local storage. This ensures security and reduces the risk of unauthorized access.

The token should only be granted read-only permissions and limited access. This ensures that it can only be used for viewing content and not for making any modifications or accessing sensitive information.

Instructions:
1- Access the link here.
2- Generate a new token.
3- Set the Repository access to "All Repositories".
4- In the Permissions section, grant only the "Read" permission for contents, leaving all other permissions as "No Access".
5- Once generated, copy the token.
6- Save the token into the input field

the link to add/change the token https://tests.stockfishchess.org/user

Also this has many optimizations and bug fixes.