Do not flag a worker as a duplicate of itself.
It may happen that a worker fails to update a task because
of server issues. It will retry 5 times and then ask for a new
task. But the server will not have marked the original task
as inactive. So before this PR the worker will be flagged as
a duplicate of itself and will be denied a new task
for some time.

In this PR we simply mark the original task as inactive
and continue.