Smarten Live Elo
This avoids fetching each 20s tabs that are not focused/visible by the user, by making one tab update at a time for multiple Live Elo tabs.
on the other hand updates immediately the new tab if it becomes the one in focus.
Allowing for less load from the API with a lazy load approach.