Proof of concept implementation of get_run via api.
With this PR /api/serialize_run hosted by the main instance can
be invoked by a secondary instance to receive a copy of the run
from the cache.

In this way there is never an issue that the run might not
be up to date.

Note: /api/serialize_run restricts access to localhost over http.
Some form of access control is necessary.