Do not evict a run from the cache which has active tasks.
A dead task (a task which has not been updated for 360s) will only
be scavenged if it belongs to a run in the cache. However clean
runs are evicted from the cache after 300s. So if the run does
not get new tasks, for example because it has low priority, the
dead tasks remain unscavenged until somebody explicitly looks at
the run (this pulls the run back in the cache).

Obviously it is not nice that dead tasks are not scavenged as soon
as possible. Moreover the fact that the behaviour of the server
is affected by executing a GET request is a violation of the HTTP
specs.