Bug fix for #2030
With the present code a finished run with run["cores"]>0 will
never be evicted from the cache. This is not a legal
condition but it happened in the past because of race conditions.

This explains why #2030 seemed to create more unvalidated runs.
The illegal runs become more and more likely to be selected for
validation as they are not evicted and eventually dominate the
cache.