Periodically validate Fishtest's internal data structures.
In this way we guarantee that the schemas in schemas.py are
kept up-to-date and hence preserve their documentary value.

Also: switch to string keys for the worker_runs dict. This is
what we do elsewhere and it makes the schema more readable:

worker_runs_schema = {
    uuid: {
        run_id: True,
        "last_run": run_id,
    }
}

Also: retire the separate variable "purge_count" and include
it in the "active_run" dict.

Requires upgrade of vtjson for the set datatype used in
"unfinished_runs_schema".