Fix missing threads option (#4)
Running the worker on macOS gives the error:

Exception running games:
Cutechess-cli says: "Warning: Base-015b2c1af2 doesn't have option Threads"

This occurs because the worker always sets the Threads option even if it's not provided in the test, but monty doesn't support Threads. Thus we work around this error until the Threads option is implemented in monty.