Different stale detection to avoid >1 workers in directory.
We touch the lock file every two seconds and check the age
of the lock file when we start the worker. If the lock
file is too old then we consider it as stale.

For the actual lock we open the lock file with the flags
O_EXCL|O_CREAT. This is atomic and will fail if the lock file
already exists.

The logic has been refactored and is contained in a separate
package packages.openlock.

Not tested on Windows but we don't use any windows specific
apis.