Add priority field to cache.
If we create a new task then we buffer the run with priority 1.
In that way it will be the first in line for saving to the db.

There are several reasons for this PR.

(1) upload_pgn is served from a secondary instance. So we want
to write the corresponding task to the db as quickly as possible,
otherwise the api will fail.

(2) If the tasks on the test page are served from a secondary
instance then there may be entries in the event log that refer to
tasks that have not yet been written to disk, probably baffling
the users. So we want to make this interval as short as possible.