Include compiler and env in engine name. Retire clear_binaries.
Currently the worker deletes cached binaries on restart. The
reason is that the user may have changed compilers, so that new
engine binaries would be matched against old engine binaries,
built with a different compiler.

In this PR we add compiler+version and a hash of selected
environment variables to the engine name so that
there is no risk of matching engine binaries built with
different compilers, or with  different environment variables.

It makes little sense to hash PATH and TEMP but they should be
included in the sandboxed environment used by the make command.

A possible enhancement (not in this PR) would be to also add
the full build command to the engine name so that engine binaries
could even be preserved across worker updates (a worker update
may change the build command).

Raise worker version to 251 (also server side)