Index worker_runs data structure by short_worker_name.
If the worker runs out of GitHub api call then the server
consults the worker_runs data structure to send the worker
only runs it has seen before.

However currently this data structure is indexed by uuid
which does not persist across worker restarts. In this PR
we change the index to the worker name, which is persistent.