Synchronize flush_all().
The former comment about deadlock seems outdated to me. Signal
handlers run in the main thread and Fishtest executes no code
in the main thread once initialization is finished. So
signal handlers behave like asynchronous events in the same way
as api calls do, and the latter appear currently to be deadlock free.