feat(server): save network file w/ Pyramid
The VPS now hosts both the fishtest server and the neural networks repository.
Save directly the network file via Pyramid retiring the FastAPI server.
Add helper script to verify and to backup to AWS S3 the network files.
The helper is scheduled to run on cron with a limited environment, it reads
the necessary environment variables directly by the .profile.