Put SPSA code in a separate file.
SPSA code related code is sprinkled at various places. It would
be beneficial to put it in a single place.

In this PR we move the SPSA code in rundb.py to a separate file.
Other code will follow in the future. We submit this PR now
because it is difficult to continuously rebase a PR that moves code.