Do not wrap byte arrays with Binary.
The Binary wrapper (with subtype 0) seems to stem from the
days that Python made no distinction between strings
and byte arrays. I checked that wrapping a byte array
with Binary makes no difference in the resulting bson
encoding. Moreover a Binary wrapper is directly decoded
as bytes and not as a Binary wrapper.

Solves:

https://tests.stockfishchess.org/actions?max_actions=1&action=log_message&user=&text=&before=1740938754.68217&run_id=