Salvage engine binaries compiled with ARCH=native
(PR on top of #2275)

Salvaging such binaries is completely safe. To mark an
engine as non-natively compiled (old stockfishes) we append
"-old" to the name of the binary.

To achieve this we refactor "setup_engine" and simplify its use.

Also: we delete the left over files with names results-xx...xx.pgn
on worker update.