RFC: Replace print by the Python logging framework in worker.
Some advantages:

(1) The Python logging framework is is thread safe so we do not
have to worry about printing from different threads.

(2) The Python logging framework is very configurable which
may be useful in the future.

This implementation is very rudimentary since I want to make sure
that other people also agree that this is a good thing to do.