Make base is not master warning more precise.
We also slightly adapt the meaning of base_same_as_master. It
now means:

The base branch is not the latest common ancestor of the
test branch and the master branch.

Also: fix diff url to compare with the official master and
not with the master of the test repo.