One more fix in dealing with #2309
See https://github.com/official-stockfish/fishtest/pull/2310#issuecomment-2961779417

A Stockfish developer can change their GitHub username. The old
url of the developer fork continues to work and is forwarded to
the new url.

However the old username can no longer be used to refer to
commits in the fork and this leads to bugs.

In this PR we replace the outdated repo url with the most current one.