Temporarily disable some javascript code.
When base is called master and master in the dev repo has
diverged from the official master then an extra button
is shown. See here

https://tests.stockfishchess.org/tests/view/6857414fc830e28d5b3937af

I think with the retirement of run["args"]["offcial_master_sha"]
this code is no longer correct. The two dot comparison should
be replaced by a three dot comparison with the official master.
However I have currently no time to study this code. It is quite
complicated.

Note that the information shown by the button is already available
from the link shown in the warning on the right.